feat Added Session in Logout and Ideal Timeout after 20 Minutes

+++ backend/routes/delete-Admin.js
@@ -1,6 +1,15 @@
 const express = require("express");
 const deleteRoutes = express.Router()  
 const signup_data = require("../models/login_signup"); 
+const session = require('express-session');
+
+
+deleteRoutes.use(session({
+  secret: 'my-session-key',
+  resave: false,
+  saveUninitialized: true
+}));
+
 
 
 
@@ -9,6 +18,12 @@ const signup_data = require("../models/login_signup");
 deleteRoutes.delete('/logindata/:id', async (req, res) => {
     try {
       const userId = req.params.id;
+      
+      // Check if the user is authenticated using session data
+      if (!req.session.token) {
+        return res.status(401).json({ message: 'Unauthorized' });
+      }
+
       const deletedUser = await signup_data.findByIdAndDelete(userId);
       if (!deletedUser) {
         return res.status(404).json({ message: 'User not found' });
@@ -23,3 +38,6 @@ deleteRoutes.delete('/logindata/:id', async (req, res) => {
   });
   
   module.exports = deleteRoutes;
+
+
+
